el commit contiene:                      parser testeado y funcionando (se realizaron comentarios en los cambios necesarios para la version inicial en las clases intervinientes

+++ src/main/kotlin/com/thegrid/communication/model/Block.kt
@@ -4,7 +4,7 @@ package com.thegrid.communication.model
  * Created by Surakituaka on 01/08/2016.
  */
 
-data class Block(val id: String, val street: Street, val length: Double, val entryNode: Node, val egressNode: Node) {
+data class Block(val id: String, val street: Street, val length: Int/*Double*/, val entryNode: Node, val egressNode: Node) {
 
     val carCapacity: Int = 3 //TODO calcular segun length
     var backStraightCarAmount: Int = 0
